Yii-notifications improvements
