+++ src/channels/EmailChannel.php
@@ -59,7 +59,7 @@ class EmailChannel extends Channel
         }
         Yii::debug("Sending email to {$this->message['to']}", __METHOD__);
         $this->mailer->getView()->params['language'] = $notification->language;
-        $message = $this->mailer->compose('@vendor/webzop/yii2-notifications/src/views/default/mail', [
+        $message = $this->mailer->compose('@vendor/webzop/yii2-notifications/src/views/notification/mail', [
             'content' => (string)$notification->getDescription(),
             'language' => $notification->language,
         ]);
